rework Overlay
Rework Overlay class so that it no longer takes the EventPropagator as an argument. It is now strictly a member of EventPropagator, not both classes members of each other.

Overlay also now creates and discards its components rather than keeps them in memory. This (hopefully) straightens out the logic of Overlay a little bit.