add mid tick actions
mid tick actions added for more precise input times. Tick times can be up to 25 ms, mid tick actions may occur at any time during the tick
if a mid tick action occurs, the player undergoes an "action tick" rather than a regular tick. They are mutually exclusive
an action tick progresses the game to the time of the action, then performs the action (either a slash or a slashdash) and return a hitbox.
    the rest of the tick occurs afterwards, where brick scroll progression is completed.
    during an action tick, the player does not move between the time of the action and the end of the tick (the action is all the player performs during that tick)
the precise time of the action is 6ms after the initial slash or dash input in a slash or slashdash command (the 6ms delay allows slash and dash to be pushed "together")